+++ ntex/src/web/types/payload.rs
@@ -208,9 +208,7 @@ impl<Err: ErrorRenderer> FromRequest<Err> for String {
         };
 
         // check content-type
-        if let Err(e) = cfg.check_mimetype(req) {
-            return Err(e);
-        }
+        cfg.check_mimetype(req)?;
 
         // check charset
         let encoding = match req.encoding() {
@@ -222,16 +220,17 @@ impl<Err: ErrorRenderer> FromRequest<Err> for String {
 
         if encoding == UTF_8 {
             Ok(str::from_utf8(body.as_ref())
-                .map_err(|_| PayloadError::Decoding)?
-                .to_owned())
+               .map_err(|_| PayloadError::Decoding)?
+               .to_owned())
         } else {
             Ok(encoding
-                .decode_without_bom_handling_and_without_replacement(&body)
-                .map(|s| s.into_owned())
-                .ok_or(PayloadError::Decoding)?)
+               .decode_without_bom_handling_and_without_replacement(&body)
+               .map(|s| s.into_owned())
+               .ok_or(PayloadError::Decoding)?)
         }
     }
 }
+
 /// Payload configuration for request's payload.
 #[derive(Clone, Debug)]
 pub struct PayloadConfig {