+++ ntex/CHANGES.md
@@ -6,6 +6,8 @@
 
 * Refactor io tls filters
 
+* Update cookie related api
+
 ## [1.0.0-b.0] - 2024-01-07
 
 * Use "async fn" in trait for Service definition